Refactor: Moved Material Components to Common
* updated Files for the same i.e. imports
