+++ src/Components/Modals/PacketNotes.js
@@ -5,7 +5,7 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import Modal from '../Modal';
+import Modal from './../Common/Modal';
 
 const PacketNotes = (props) => {
