+++ src/Redux/Actions/HexCode.js
@@ -1,4 +1,4 @@
-import TYPE from "../Types/HexCode";
+import TYPE from "./../Types/HexCode";
 
 export const setHexCode = (data) => (dispatch) => {
   dispatch({ type: TYPE.SET_HEX_CODE, payload: data });
