+++ src/Redux/Actions/Proxy.js
@@ -1,8 +1,8 @@
-import TYPE from "../Types/Proxy";
-import TYPE_PAGE from "../Types/Page";
+import TYPE from "./../Types/Proxy";
+import TYPE_PAGE from "./../Types/Page";
 
-import { packets } from "../../__Mocks__/Data/Packets";
-import { connections } from "../../__Mocks__/Data/Connection";
+import { packets } from "./../../__Mocks__/Data/Packets";
+import { connections } from "./../../__Mocks__/Data/Connection";
 
 export const getPackets = (data) => (dispatch, getState) => {
   dispatch({ type: TYPE_PAGE.SHOW_BACKDROP, payload: { count: ++getState().page.backdrop.count } });
